Add SV variables to DiJet variables
Requires re-fitting of MiniAOD jets to get SV info.
Then, the BoostedAnalyzer will have to read this new
jet collection instead of the default slimmedJets. Started
to adapt BoostedAnalyzer to be configurable for input collections
but this is not yet consistent throughout the collections.
Defaults should not be touched except for DiJet analysis.